requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
pycodestyle==2.9.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.11
zipp==3.8.1